Including legend into exported image.
Scaling the legend I'm trying to keep it close to proportion of the image on the screen.
Since currently pixel ratio for 3D scene if fixed and not adjustable in 3D legend always has
constant size (but actual code takes pixel ratio into account).

in 2D mode scale of legend depends on current scale: if you apply to the exported image the same
scale as it's on the screen then the legend will have the same size as on the sceen. User may not
see where exactly legend will be placed on the exported picture but she may see relative sizes of
objects on the image and the lagend.

There is a trick: because logical pixels depend on the page scale it's possible to increase size of exported image
by zooming out (with Ctrl-'-').

Issue: #1