Removing a copy of mesh verteces.
Copy of verteces copied before centring to preserve source values (in common basis with spots coordinates).

Now geometry coordinates are preserved in their initial values. The same transformation applied on the fly (on GPU).
To achieve that introduced a new field in Scene3D: _meshContainer. The mesh inside the mesh container is moved in such
a way that makes center of the mesh container to be a center of bounding rectangle of the mesh.

Since gemoetry coordinates are now in the same basis as spots coordinates then worldMatrix of the mesh
could be used to translate spot coordinates to the world basis.

Issue: #13.